support splitting into files
