+++ kotlinx.fuzz.api/src/main/kotlin/kotlinx/fuzz/KFuzzConfig.kt
@@ -10,162 +10,6 @@ import kotlin.reflect.jvm.isAccessible
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 
-
-interface KFConfig {
-    val global: GlobalConfig
-    val target: TargetConfig
-
-    fun toPropertiesMap(): Map<String, String>
-}
-
-interface GlobalConfig {
-    val workDir: Path
-}
-
-interface TargetConfig {
-    val maxTime: Duration
-}
-
-class KFBuilder(
-    private val propertiesMap: Map<String, String>,
-) {
-    // =========== config declarations ===========
-
-    inner class GlobalConfigImpl internal constructor() : GlobalConfig {
-        override var workDir by Prop<Path>(
-            nameSuffix = "workDir",
-            fromString = { Path.of(it) },
-            intoString = { it.absolutePathString() },
-        )
-    }
-
-    inner class TargetConfigImpl internal constructor() : TargetConfig {
-        override var maxTime by Prop<Duration>(
-            nameSuffix = "maxTime",
-            fromString = { Duration.parse(it) },
-            intoString = { it.toString() },
-            validate = { require(it.isPositive()) { "maxTime must be positive!" } },
-            default = { Duration.parse("1m") }
-        )
-    }
-
-    // =========== builder implementation ===========
-
-    private var isBuilt = false
-    private var editOverride = false
-    private val delegates = mutableListOf<Prop<*>>()
-
-    /**
-     * A fuzzing config property. The value is looked up with the following priority:
-     *
-     * 1) editOverride
-     * 2) property map
-     * 3) editFallback
-     * 4) default
-     */
-    private inner class Prop<T : Any>(
-        nameSuffix: String,
-        private val fromString: (String) -> T,
-        private val intoString: (T) -> String,
-        private val validate: (T) -> Unit = {}, // throws
-        private val default: (() -> T)? = null,
-    ) {
-        val name = CONFIG_NAME_PREFIX + nameSuffix
-
-        init {
-            delegates += this
-        }
-
-        private var value: T? = null
-
-        operator fun getValue(thisRef: Any?, property: KProperty<*>): T {
-            check(isBuilt) { "cannot get value, config is not built yet!" }
-            return value!!
-        }
-
-        operator fun setValue(thisRef: Any?, property: KProperty<*>, value: T) {
-            check(!isBuilt) { "cannot set value, config is already built!" }
-            this.value = value
-        }
-
-        val isSet get() = value != null
-
-        fun setFromPropertiesMap() {
-            propertiesMap[name]?.let {
-                value = fromString(it)
-            }
-        }
-
-        fun setFromDefault() {
-            default?.let {
-                value = it()
-            }
-        }
-
-        fun getStringValue(): String {
-            check(isBuilt) { "cannot get string value, config is not built yet!" }
-            return intoString(value!!)
-        }
-
-        fun validate() {
-            check(value != null) { "property '$name' was not set!" }
-            validate(value!!)
-        }
-    }
-
-    inner class KFConfigImpl internal constructor() : KFConfig {
-        override val global = GlobalConfigImpl()
-        override val target = TargetConfigImpl()
-
-        override fun toPropertiesMap(): Map<String, String> {
-            check(isBuilt) { "cannot get properties map, config is not built yet!" }
-            return delegates.associate { it.name to it.getStringValue() }
-        }
-    }
-
-    companion object {
-        fun fromSystemProperties(): KFBuilder = KFBuilder(getSystemPropertiesMap())
-        fun fromAnotherConfig(config: KFConfig): KFBuilder = KFBuilder(config.toPropertiesMap())
-
-        val CONFIG_NAME_PREFIX = "kotlinx.fuzz."
-    }
-
-    private val overrideSteps = mutableListOf<KFConfigImpl.() -> Unit>()
-    private val fallbackSteps = mutableListOf<KFConfigImpl.() -> Unit>()
-
-    fun editOverride(editor: KFConfigImpl.() -> Unit): KFBuilder = this.also {
-        overrideSteps.add(editor)
-    }
-
-    fun editFallback(editor: KFConfigImpl.() -> Unit): KFBuilder = this.also {
-        fallbackSteps.add(editor)
-    }
-
-    fun build(): KFConfig {
-        check(!isBuilt) { "config is already built!" }
-        try {
-            val config = KFConfigImpl()
-            /*
-              To guarantee the priority in the following order...
-              1) editOverride
-              2) property map
-              3) editFallback
-              4) default
-              ...we can set values in backwards order.
-             */
-            delegates.forEach { it.setFromDefault() }
-            fallbackSteps.forEach { it.invoke(config) }
-            delegates.forEach { it.setFromPropertiesMap() }
-            overrideSteps.forEach { it.invoke(config) }
-
-            isBuilt = true
-            return config
-        } catch (e: Throwable) {
-            throw ConfigurationException(e)
-        }
-    }
-}
-
 /**
  * Class that stores generals fuzzing configuration
  *
@@ -401,15 +245,5 @@ private inline fun <T : KFuzzConfig> wrapConfigErrors(buildConfig: () -> T): T =
     throw when (e) {
         is ConfigurationException -> e
         else -> ConfigurationException(e)
-    }
-}
-
-fun getSystemPropertiesMap(): Map<String, String> = buildMap {
-    val properties = System.getProperties()
-    val propNames = properties.propertyNames()
-    for (name in propNames) {
-        val key = name.toString()
-        val value = properties.getProperty(key).toString()
-        put(key, value)
     }
 }
