+++ kotlinx.fuzz.gradle/src/main/kotlin/kotlinx/fuzz/gradle/FuzzConfigDSL.kt
@@ -1,0 +1,13 @@
+package kotlinx.fuzz.gradle
+
+interface FuzzConfigDSL {
+    // global
+
+    // target (on same level)
+
+    // engine
+
+    // jacoco
+
+}
+
